requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==4.2.1
djangorestframework==3.16.0
filelock==3.18.0
identify==2.6.12
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
virtualenv==20.31.2
